forgot this in last commit

+++ crypto/sha.h
@@ -5,19 +5,26 @@
 
 /* each returns a byte array interpreted as a BE-int */
 
-void bksmt_sha1(unsigned char *, int, unsigned char[20]);
+/* Args: msg, len, output */
+void bksmt_sha1(unsigned char *, uint64_t, unsigned char[20]);
 
-void bksmt_sha256(unsigned char *, int, unsigned char[32]);
+/* Args: msg, len, output */
+void bksmt_sha256(unsigned char *, uint64_t, unsigned char[32]);
                                                          
-void bksmt_sha224(unsigned char *, int, unsigned char[28]);
+/* Args: msg, len, output */
+void bksmt_sha224(unsigned char *, uint64_t, unsigned char[28]);
                                                          
-void bksmt_sha512(unsigned char *, int, unsigned char[64]);
+/* Args: msg, len, output */
+void bksmt_sha512(unsigned char *, uint64_t, unsigned char[64]);
                                                          
-void bksmt_sha384(unsigned char *, int, unsigned char[48]);
+/* Args: msg, len, output */
+void bksmt_sha384(unsigned char *, uint64_t, unsigned char[48]);
 
-void bksmt_sha512t224(unsigned char *, int, unsigned char[28]);
+/* Args: msg, len, output */
+void bksmt_sha512t224(unsigned char *, uint64_t, unsigned char[28]);
                                                              
-void bksmt_sha512t256(unsigned char *, int, unsigned char[32]);
+/* Args: msg, len, output */
+void bksmt_sha512t256(unsigned char *, uint64_t, unsigned char[32]);
 
 
 #endif /* __BKSMT_CRYPTO_SHA_H__ */
